requirements.txt
nltk==3.9.1
ollama==0.3.3
openai==1.46.1
pandas==2.2.2
python-dotenv==1.0.1
python_Levenshtein==0.26.0
rich==13.8.1
rouge_score==0.1.2
openpyxl
h2
pyyaml
